Moved HttpResponseBuilder out of HttpResponse. Added ResponseData. This initializes HttpResponses.
