+++ frontend/jobapptrackerfrontend/src/model/interfaces/init/ResponseData.ts
@@ -1,0 +1,6 @@
+
+export interface HttpResponseData<T> {
+    data: T,
+    status: number,
+    errorMessage: string
+};
